+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # ChangeLog
 
+## 1.4.7 (2018/04/03)
+ * Support async functions (thanks to @hoegertn)
+
+## 1.4.6 (2018/03/08)
+ * Fix region detection
+ * Add --envdestroy option to restore the environment after testing.
+ * Support builtin Error object handling
+ * More examples & better README
+
+## 1.4.5 (2017/11/15)
+ * Better environment variables managing
+ * Fix lambda ending rules (context.done/fail/succeed)
+
 ## 1.4.4 (2017/10/24)
  * Timeout fix
  * Better context management